feat(xo-server): implement rolling pull reboot
