+++ packages/xo-server/src/xo-mixins/authorization.mjs
@@ -33,6 +33,7 @@ const AUTHORIZATIONS = {
   LIST_MISSING_PATCHES: STARTER,
   POOL_EMERGENCY_SHUTDOWN: ENTERPRISE,
   ROLLING_POOL_UPDATE: ENTERPRISE,
+  ROLLING_POOL_REBOOT: ENTERPRISE,
 }
 
 export default class Authorization {
