+++ packages/xo-server/src/xo-mixins/xen-servers.mjs
@@ -641,7 +641,7 @@ export default class XenServers {
 
   async rollingPoolUpdate($defer, pool) {
     const app = this._app
-
+    await app.checkFeatureAuthorization('ROLLING_POOL_UPDATE')
     const [schedules, jobs] = await Promise.all([app.getAllSchedules(), app.getAllJobs('backup')])
 
     const poolId = pool.id
